refactor(store): 优化存储模块代码结构和可读性
- 将 CommitLogAppendHandler 中的 messageStoreConfig 属性设为 final
- 更新 CommitLogFileModel 中的 MMap 工具方法调用为读写模式
- 使用 Lombok @Getter 注解简化 CommitLogFileModel 的 getter 方法
- 修改 MMapUtil 工具类以支持更安全的文件映射创建逻辑
- 修复 StorageUnits 中 TERABYTE 类型定义不一致的问题
- 移除 BrokerStartup 中 initConfigContext 方法不必要的返回值- 清理测试类中冗余的异常声明和硬编码路径配置

+++ catmq-store/src/main/java/com/aoaojiao/catmq/store/constants/StorageUnits.java
@@ -8,5 +8,5 @@ public class StorageUnits {
     public static final int KILOBYTE = 1024;
     public static final int MEGABYTE = KILOBYTE * 1024;
     public static final int GIGABYTE = MEGABYTE * 1024;
-    public static final int TERABYTE = GIGABYTE * 1024;
+    public static final long TERABYTE = GIGABYTE * 1024L;
 }
